core-scroll: Allow small movements while clicking. (#101)
* Allow small movements while clicking.

When the user clicks on an element and the cursor moves slightly, it is most
likely that the user wanted to click in stead of drag.

Currently, as reported by users, they have a hard time clicking on elements
inside a element controlled by core-scroll. It is hard to know exactly why, but
if they are moving their cursor, even 1px while clicking on an element the
click is prevented. This commit addresses that issue.

* Move consts to top